pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "emoji>=2.15.0",
    "google-genai>=1.2.0",
    "httpx==0.27.2",
    "openai==1.12.0",
    "pillow==10.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv==1.0.1",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
